requirements.txt
pytest
pony
selenium
jsonpickle
pymysql
mysql-connector-python
pytest_allure_adaptor-1.7.1-py2-none-any.whl

